Add _user_from_userid helper to handle user names with commas
Some users have commas in their names, which causes errors because
commas are also used in Merlin to separate the user's name from the
group name. For example "Somebody, N, Guests" would become "Somebody"
in group "N" instead of "Somebody, N" in group "Guests". This
helper function takes the last item as the group and concatenates
the rest to form the name